Simplify base image value object and removal of twitter image value object
